Refactored the Doctrine implementation to use the same class
