+++ Entity/GroupManager.php
@@ -13,67 +13,19 @@ namespace FOS\UserBundle\Entity;
 
 use FOS\UserBundle\Model\GroupInterface;
 use Doctrine\ORM\EntityManager;
-use FOS\UserBundle\Model\GroupManager as BaseGroupManager;
+use FOS\UserBundle\Doctrine\GroupManager as BaseGroupManager;
 
+/**
+ * BC class for people extending it in their bundle.
+ * TODO Remove this class on July 31st
+ */
 class GroupManager extends BaseGroupManager
 {
     protected $em;
-    protected $class;
-    protected $repository;
 
     public function __construct(EntityManager $em, $class)
     {
+        parent::__construct($em, $class);
         $this->em = $em;
-        $this->repository = $em->getRepository($class);
-
-        $metadata = $em->getClassMetadata($class);
-        $this->class = $metadata->name;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function deleteGroup(GroupInterface $group)
-    {
-        $this->em->remove($group);
-        $this->em->flush();
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function getClass()
-    {
-        return $this->class;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function findGroupBy(array $criteria)
-    {
-        return $this->repository->findOneBy($criteria);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function findGroups()
-    {
-        return $this->repository->findAll();
-    }
-
-    /**
-     * Updates a group
-     *
-     * @param GroupInterface $group
-     * @param Boolean        $andFlush Whether to flush the changes (default true)
-     */
-    public function updateGroup(GroupInterface $group, $andFlush = true)
-    {
-        $this->em->persist($group);
-        if ($andFlush) {
-            $this->em->flush();
-        }
     }
 }
