+++ Entity/UserManager.php
@@ -13,91 +13,22 @@ namespace FOS\UserBundle\Entity;
 
 use Doctrine\ORM\EntityManager;
 use FOS\UserBundle\Model\UserInterface;
-use FOS\UserBundle\Model\UserManager as BaseUserManager;
+use FOS\UserBundle\Doctrine\UserManager as BaseUserManager;
 use FOS\UserBundle\Util\CanonicalizerInterface;
 use Symfony\Component\Security\Core\Encoder\EncoderFactoryInterface;
 
+/**
+ * BC class for people extending it in their bundle.
+ * TODO Remove this class on July 31st
+ */
 class UserManager extends BaseUserManager
 {
     protected $em;
-    protected $class;
-    protected $repository;
 
-    /**
-     * Constructor.
-     *
-     * @param EncoderFactoryInterface $encoderFactory
-     * @param CanonicalizerInterface  $usernameCanonicalizer
-     * @param CanonicalizerInterface  $emailCanonicalizer
-     * @param EntityManager           $em
-     * @param string                  $class
-     */
     public function __construct(EncoderFactoryInterface $encoderFactory, CanonicalizerInterface $usernameCanonicalizer, CanonicalizerInterface $emailCanonicalizer, EntityManager $em, $class)
     {
-        parent::__construct($encoderFactory, $usernameCanonicalizer, $emailCanonicalizer);
+        parent::__construct($encoderFactory, $usernameCanonicalizer, $emailCanonicalizer, $em, $class);
 
         $this->em = $em;
-        $this->repository = $em->getRepository($class);
-
-        $metadata = $em->getClassMetadata($class);
-        $this->class = $metadata->name;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function deleteUser(UserInterface $user)
-    {
-        $this->em->remove($user);
-        $this->em->flush();
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function getClass()
-    {
-        return $this->class;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function findUserBy(array $criteria)
-    {
-        return $this->repository->findOneBy($criteria);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function findUsers()
-    {
-        return $this->repository->findAll();
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function reloadUser(UserInterface $user)
-    {
-        $this->em->refresh($user);
-    }
-
-    /**
-     * Updates a user.
-     *
-     * @param UserInterface $user
-     * @param Boolean       $andFlush Whether to flush the changes (default true)
-     */
-    public function updateUser(UserInterface $user, $andFlush = true)
-    {
-        $this->updateCanonicalFields($user);
-        $this->updatePassword($user);
-
-        $this->em->persist($user);
-        if ($andFlush) {
-            $this->em->flush();
-        }
     }
 }
