+++ Resources/config/couchdb.xml
@@ -5,7 +5,7 @@
     xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
 
     <parameters>
-        <parameter key="fos_user.user_manager.class">FOS\UserBundle\CouchDocument\UserManager</parameter>
+        <parameter key="fos_user.user_manager.class">FOS\UserBundle\Doctrine\UserManager</parameter>
     </parameters>
 
     <services>
