+++ Resources/config/mongodb.xml
@@ -5,7 +5,7 @@
     xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
 
     <services>
-        <service id="fos_user.user_manager.default" class="FOS\UserBundle\Document\UserManager" public="false">
+        <service id="fos_user.user_manager.default" class="FOS\UserBundle\Doctrine\UserManager" public="false">
             <argument type="service" id="security.encoder_factory" />
             <argument type="service" id="fos_user.util.username_canonicalizer" />
             <argument type="service" id="fos_user.util.email_canonicalizer" />
