+++ Resources/config/orm_group.xml
@@ -5,7 +5,7 @@
     xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
 
     <parameters>
-        <parameter key="fos_user.group_manager.class">FOS\UserBundle\Entity\GroupManager</parameter>
+        <parameter key="fos_user.group_manager.class">FOS\UserBundle\Doctrine\GroupManager</parameter>
     </parameters>
 
     <services>
